Another draft of Builder
