+++ src/Builder/ParamBuilder.php
@@ -3,10 +3,10 @@ declare(strict_types=1);
 namespace Feather\Builder;
 
 class ParamBuilder extends \Feather\Base {
-    private $values;
-    private $paramsCount = 0;
+    protected $values;
+    protected $paramsCount = 0;
 
-    const PREFIX = ':param';
+    const PREFIX = ':p';
 
     /**
      * Return :paramNUMBER 
@@ -18,5 +18,18 @@ class ParamBuilder extends \Feather\Base {
         $param = self::PREFIX.$this->paramsCount++;
         $this->values[$param] = $value;
         return $param;
+    }
+
+    public function clearparams() {
+        $this->values = [];
+        $this->paramsCount = 0;
+    }
+
+    public function getParams() {
+        return $this->value;
+    }
+
+    public function getCount() {
+        return $this->paramsCount;
     }
 }
