+++ src/Builder/QueryBuilder.php
@@ -1,0 +1,7 @@
+<?php
+declare(strict_types=1);
+namespace Feather\Builder;
+
+class QueryBuilder extends \Feather\Base {
+
+}
