+++ src/Exceptions/DriverNotSetException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Feather\Exceptions;
+
+class DriverNotSetException extends DbException {
+
+}
