+++ src/Statement.php
@@ -1,0 +1,7 @@
+<?php
+declare(strict_types=1);
+namespace Feather;
+
+class Statement extends Base {
+
+}
