+++ tests/MySQLDriverTest.php
@@ -1,0 +1,11 @@
+<?php 
+
+use \Feather\Driver\MySQL\MySQL;
+
+class MySQLDriverTest extends \Test\DriverBase
+{
+    public function __construct() {
+        $this->driver = new MySQL();
+        parent::__construct();
+    }
+}
